[refactor] confirm image silde component 분리
